+++ src/main/java/lotto/view/OutputView.java
@@ -1,8 +1,10 @@
-package lotto;
+package lotto.view;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import lotto.model.Lotto;
+import lotto.model.Rank;
 
 public class OutputView {
